+++ examples/armbench-segmentation/tensorleap.py
@@ -14,7 +14,7 @@ from code_loader.contract.enums import (
 from code_loader.contract.datasetclasses import PreprocessResponse
 from pycocotools.coco import COCO
 
-from armbench_segmentation.config_utils import CONFIG
+from armbench_segmentation.config import CONFIG
 from armbench_segmentation.gcs_utils import _download
 from armbench_segmentation.preprocessing import load_set
 from armbench_segmentation.utils.general_utils import count_obj_masks_occlusions, \
